barrier: fix ABT_barrier_free
ABT_barrier might be freed right after ABT_barrier_wait().  A lock is taken in
ABT_barrier_free() in order not to free ABT_barrier() that is being used.  The
number of waiters, however, is checked before this lock, which causes an error
if assert() is enabled; before taking the lock, there might exist some waiters.
This patch fixes it.