sched_config: fix the declaration of ABT_sched_basic_freq
ABT_sched_basic_freq is declared as a 0-idx configuration variable, but idx >= 0
is preserved for parameters of a user-defined scheduler.  This patch moves its
declaration to config.c and changes its idx to -4.  Note that it is declared as
an external variable, so changing its value does not break the ABI.

ABT_sched_basic_wait_freq is not used, so it's removed.