test/basic: fix sched_prio
verify_exec_order() of the primary execution stream can be disturbed by the
primary ULT.  Specifically, the primary ULT can be pushed to its pool (= the
highest priority pool) during verify_exec_order() if it is resumed by
another execution stream, failing verify_exec_order() since there is a ULT in
a high-priority pool but a ULT associated with a low-priority pool is scheduled
first.

This patch fixes it by changing the primary ULT's pool to the lowest priority
pool.  Note that this depends on the timing of a resume operation and is very
rare (e.g., 0.1% on my laptop).