mem: implement a stack canary-based overflow detection mechanism
A stack canary is a special value that is written at the bottom of the ULT
stack.  The Argobots runtime can detect if this value is overwritten by stack
overflow when a stack is freed.  This mechanism does not catch stack overflow
when it happens, but this is a quick method to check if there is no stack
overflow.

This option is disabled by default.  It can be enabled by setting
--enable-stack-overflow-check=canary-XX where XX is the canary size.