global: clarify functions that use p_global
To clarify functions that must be called when Argobots is initialized, all the
internal functions take p_global as an argument if they need to access it.  The
top-level functions (e.g., ABT_xxx()) must get p_global using wrapper functions.
Developers can easily understand whether a top-level function is really
independent of the Argobots initialization state; a function that calls a
function taking p_global is a function that needs Argobots initialization.

This change might badly affect the compiler optimization, but the performance
impact should be little.