self: create self functions that get and set information about self
Functions that get and set properties of the underlying work units and execution
streams should be in the "self" category, but some of the functions belong to
different categories, which is confusing.  This patch creates functions with
proper names as follows:
 - ABT_xstream_self      -> ABT_self_get_xstream
 - ABT_xstream_self_rank -> ABT_self_get_xstream_rank
 - ABT_thread_self       -> ABT_self_get_thread
 - ABT_thread_self_id    -> ABT_self_get_thread_id
 - ABT_task_self         -> ABT_self_get_task
 - ABT_task_self_id      -> ABT_self_get_task_id
 - ABT_key_set           -> ABT_self_set_specific
 - ABT_key_get           -> ABT_self_get_specific

Note that this patch does not remove the existing functions, so the ABI/API
compatibility will be kept.