self: create functions that controlling self ESs and ULTs
Functions that change the control of the underlying work units and execution
streams should be in the "self" category (e.g., ABT_self_suspend()), but some of
the functions belong to different categories, which is confusing.  This patch
creates the following functions:
 - ABT_thread_yield -> ABT_self_yield
 - ABT_thread_exit  -> ABT_self_exit

Note that this patch does not remove the existing functions, so the ABI/API
compatibility will be kept.

This patch does not create aliasesv for ABT_thread_yield_to() and
ABT_xstream_exit() because the behaviors of those functions are unclear.
Improved versions of these functions should be provided in the future.