test/util: relax affinity of the primary execution stream
New ext_thread tests create Pthreads on the primary execution stream.
On Linux, a newly created Pthread will inherit its Parent's affinity
setting, so when affinity is enabled and no specific affinity is set
(i.e., when no ABT_SET_AFFINIY is set), the primary execution stream and
external Pthreads are pinned to a single core, which significantly
degrades the performance because the primary execution stream does not
sleep (while Pthreads internally sleep on futex or pthread_cond_t).
This patch relaxes the affinity setting of the primary execution stream
to avoid this oversubscription.