xstream: fix xstream_update_main_sched
When xstream_update_main_sched() updates the underlying execution
stream's main scheduler, the current ULT is pushed to the pool before
context switching, so if that pool is shared with other execution
streams, this ULT might be executed before saving its context, crashing
the program.  This patch fixes it by pushing this ULT after context
switch.