pool: set a special bit for ABT_unit associated with a built-in pool
To find an associated pool from ABT_unit, the future commit will
implement a unit-thread hash table.  This hash table is heavier than the
existing algorithm that ignores such mapping.  To avoid performance
degradation of most programs that use built-in pools, special
optimization for built-in pools is needed.  This patch adds a new bit to
built-in ABT_unit to prepare for such optimization in the future commit.