unit: unify unit-pool management mechanisms
The current Argobots does not check unit-pool mapping, so it could be
problematic if the user uses different types of pools and move ABT_unit
from one to another.  This patch implements a hash table to map ABT_unit
and ABTI_pool (via ABTI_thread) and updates the hash table whenever the
pool association is changed.  If the user only uses built-in pools, a
hash table is skipped to minimize the overheads.