xstream: avoid freeing sched before it finishes when update main_sched
In xstream_update_main_sched(), the caller ULT frees the main scheduler
when it updates its underlying main scheduler.  If the caller ULT is
scheduled by another scheduler, however, the scheduler that is still
used can be freed.  This patch fixes it by using a request to release a
scheduler.