env: support ABT_STACK_OVERFLOW_CHECK=mprotect(_strict)
The new mprotect-based stack overflow check does not add a large
overhead to the existing ULT mechanism if disabled.  Since it should be
useful for debugging, this patch makes the setting of the mprotect-based
stack overflow dynamically configureable.

The user can set ABT_STACK_OVERFLOW_CHECK=mprotect or mprotect_strict to
enable this feature.