global: retrieve the system page size
mprotect() needs the predefined system page size (usually 4KB).  It
should be automatically obtained by getpagesize() in most cases, but
this patch also adds a new environmental variable to accept a user value
to deal with cases where getpagesize() is undefined or mprotect() on
that system has a different restriction.

The next commit will use this value.