mem: add mprotect-based guard for stacks allocated outside mempools
This patch supports the mprotect-based guard for stacks that are
allocated without memory pools.  A few branches are added to the
fork-join path even if the stack guard is disabled, but this overhead
should be acceptable considering that non-default ULT allocation is not
performance critical.  If this overhead is unacceptable, we should
create another patch to remove those branches at configure time.