configure: support --enable-stack-unwind=unwind-only
By default, Argobots prints the raw stack data as well as the function
stack even if the stack unwinding feature is enabled.  This patch adds
a new option --enable-stack-unwind=unwind-only to disable it.

Note that Argobots does not change the current behavior by default.  The
raw stack will be printed if --enable-stack-unwind is used without an
argument.