error: implement ABTI_UB_ASSERT
ABTI_UB_ASSERT() is a debugging macro to warn incorrect use of Argobots
APIs.  This check is sometimes expensive, so it is disabled by default;
if such an incorrect input is given, undefined behavior occurs as
explained in the Argobots specification.  This check, however, would
be help debug Argobots programs.

ABTI_UB_ASSERT() is enabled when --enable-debug={ub_assert, all, most}
is set.