fcontext: introduce a new set of fcontext functions
Currently, Argobots uses a different set of unoptimized fcontext
functions for various architectures, making maintenance complicated.
This patch implements a new set of fcontext functions and uses it for
user-level context switching.

Note that it is hard to split this commit since we need to change
everything at once; otherwise Argobots is not functional.