ucontext: call getcontext in ABTD_ythread_context_(re)init
For an unknown reason, lazy getcontext() + makecontext() does not
properly handle vectorized code on POWER9 (with Clang 10.0).
See https://github.com/pmodels/argobots/pull/338 for details.

This patch fixes it by eagerly calling getcontext().  We did not
observe, but the same problem can happen on other machines, so this
change applies to all the platforms.  This should not affect the
performance critical paths since anyway ucontext is slow. On POWER9, we
recommend fcontext, which is faster (enabled by default).