info: make more information available before ABT_init
This patch updates ABT_info_query_config() so that the user can get more
information without initializing the Argobots runtime.  This is useful
especially when an application wants to examine the underlying Argobots
configuration before ABT_init().

Note that the user needs to enable an experimental Argobots 2.0 API
(--enable-ver20-api) since the 1.x specification defines that
ABT_info_query_config() should return an error when Argobots is not
initialized.