thread: introduce a callback-based request handling mechanism
Several threading algorithms need to perform some operation (e.g.,
pushing the current ULT to its associated pool) on the next ULT, which
is hard to follow since they are implemented in different files.  This
patch introduces a callback-based mechanism so that the developer can
easily understand which operation will be performed after the context
switch.  The next ULT does not need to know the previous ULT, which
also makes the logic simpler.

This patch introduces new fcontext functions to efficiently call a
callback function.