configure: apply cancellation/migration configs to all thread types
It is nonsense to enable different feature sets for ULTs (yieldable) and
tasklets (non-yieldable).  This patch unifies the configuration options.
Note that this patch does not break the backward compatibility.  It
could potentially affect the program behavior if users disable only
either ULT or tasklet cancellation--which should be a very rare case
since very few users currently use cancellation features because of its
instability.