xstream: fix ABT_xstream_set_main_sched_basic in an error case
If ABT_xstream_set_main_sched_basic() takes automatically freed pools
but it failed after scheduler creation (i.e., on setting the main
scheduler because of memory allocation failure in, for example,
ABTI_thread_set_associated_pool()), the user-given automatically freed
pools are freed when the scheduler is freed.  They should be kept alive
in such a case.  This patch fixes it.