pool_user_def: add ABT_pool_user_def to implement a user-defined pool
ABT_pool_user_def and its utility functions allow creating a
user-defined pool that uses a new pool interface.  ABT_pool_user_def is
a pointer to ABT_pool_def, so it can be passed to ABT_pool_create(),
which is for backward compatibility.  This patch also adds descriptions
to newly introduced constants and types so that the user can use this
new interface.