examples: use ABT_pool_xxx_thread and ABT_self_schedule for scheduler
The user is recommended to use ABT_pool_pop_thread() and
ABT_self_schedule() instead of ABT_pool_pop() and ABT_xstream_run_unit()
because the user does not need to be aware of ABT_unit.  This also
allows a scheduler to easily inspect work units (e.g., a thread
function or a ULT-level TLS), increasing the flexibility of scheduling.