info: add start/end tags and pack zeroes in stack dump
Presently, there is no tag to indicate when ULTs stacks
dump has completed, when this is difficult to determine
as this process occurs asynchronously. So, start and
end markers with timestamp have been added.

Also, the raw stack dump do not pack the full zeroes
parts, which sometimes causes tons of unnecessary lines
of zeroes to be printed. So, they are packed now to
improve readability.

Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>